add QCOW2 build mechanism

+++ build.sh
@@ -1,4 +1,7 @@
 #!/bin/bash -e
+
+#set -x
+
 # shellcheck disable=SC2119
 run_sub_stage()
 {
@@ -22,6 +25,11 @@ EOF
 				on_chroot << EOF
 apt-get install --no-install-recommends -y $PACKAGES
 EOF
+			    if [ "${USE_QCOW2}" = "1" ]; then
+				on_chroot << EOF
+apt-get clean
+EOF
+			    fi
 			fi
 			log "End ${SUB_STAGE_DIR}/${i}-packages-nr"
 		fi
@@ -32,6 +40,11 @@ EOF
 				on_chroot << EOF
 apt-get install -y $PACKAGES
 EOF
+			    if [ "${USE_QCOW2}" = "1" ]; then
+				on_chroot << EOF
+apt-get clean
+EOF
+			    fi
 			fi
 			log "End ${SUB_STAGE_DIR}/${i}-packages"
 		fi
@@ -82,17 +95,27 @@ EOF
 run_stage(){
 	log "Begin ${STAGE_DIR}"
 	STAGE="$(basename "${STAGE_DIR}")"
+
 	pushd "${STAGE_DIR}" > /dev/null
-	unmount "${WORK_DIR}/${STAGE}"
+
 	STAGE_WORK_DIR="${WORK_DIR}/${STAGE}"
 	ROOTFS_DIR="${STAGE_WORK_DIR}"/rootfs
+
+	if [ "${USE_QCOW2}" = "1" ]; then 
+	    if [ ! -f SKIP ]; then
+		load_qimage
+	    fi
+	else
+	    unmount "${WORK_DIR}/${STAGE}"
+	fi
+	
 	if [ ! -f SKIP_IMAGES ]; then
 		if [ -f "${STAGE_DIR}/EXPORT_IMAGE" ]; then
 			EXPORT_DIRS="${EXPORT_DIRS} ${STAGE_DIR}"
 		fi
 	fi
 	if [ ! -f SKIP ]; then
-		if [ "${CLEAN}" = "1" ]; then
+		if [ "${CLEAN}" = "1" ] && [ "${USE_QCOW2}" = "0" ] ; then
 			if [ -d "${ROOTFS_DIR}" ]; then
 				rm -rf "${ROOTFS_DIR}"
 			fi
@@ -109,7 +132,13 @@ run_stage(){
 			fi
 		done
 	fi
-	unmount "${WORK_DIR}/${STAGE}"
+
+	if [ "${USE_QCOW2}" = "1" ]; then 
+		unload_qimage
+	else
+    	    unmount "${WORK_DIR}/${STAGE}"
+	fi
+
 	PREV_STAGE="${STAGE}"
 	PREV_STAGE_DIR="${STAGE_DIR}"
 	PREV_ROOTFS_DIR="${ROOTFS_DIR}"
@@ -140,6 +169,15 @@ do
 			;;
 	esac
 done
+
+term() {
+    if [ "${USE_QCOW2}" = "1" ]; then
+	log "Unloading image"
+	unload_qimage
+    fi
+}
+
+trap term EXIT INT TERM
 
 export PI_GEN=${PI_GEN:-pi-gen}
 export PI_GEN_REPO=${PI_GEN_REPO:-https://github.com/RPi-Distro/pi-gen}
@@ -208,6 +246,10 @@ source "${SCRIPT_DIR}/common"
 # shellcheck source=scripts/dependencies_check
 source "${SCRIPT_DIR}/dependencies_check"
 
+export USE_QCOW2="${USE_QCOW2:-1}"
+export BASE_QCOW2_SIZE=${BASE_QCOW2_SIZE:-12G}
+source "${SCRIPT_DIR}/qcow2_handling"
+
 dependencies_check "${BASE_DIR}/depends"
 
 #check username is valid
@@ -237,13 +279,29 @@ for EXPORT_DIR in ${EXPORT_DIRS}; do
 	# shellcheck source=/dev/null
 	source "${EXPORT_DIR}/EXPORT_IMAGE"
 	EXPORT_ROOTFS_DIR=${WORK_DIR}/$(basename "${EXPORT_DIR}")/rootfs
-	run_stage
+        QIMAGE="image-$(basename "${EXPORT_DIR}").qcow2"
+	if [ "${USE_QCOW2}" = "1" ]; then
+	    USE_QCOW2=0
+	    mount_qimage "${WORK_DIR}/${QIMAGE}" "${EXPORT_ROOTFS_DIR}"
+	    echo "Mounting image ${WORK_DIR}/${QIMAGE} to export rootfs ${EXPORT_ROOTFS_DIR}"
+	    run_stage
+	    unload_qimage
+	    USE_QCOW2=1
+	else
+	    run_stage
+	fi 
 	if [ "${USE_QEMU}" != "1" ]; then
 		if [ -e "${EXPORT_DIR}/EXPORT_NOOBS" ]; then
 			# shellcheck source=/dev/null
 			source "${EXPORT_DIR}/EXPORT_NOOBS"
 			STAGE_DIR="${BASE_DIR}/export-noobs"
-			run_stage
+			if [ "${USE_QCOW2}" = "1" ]; then
+			    USE_QCOW2=0
+			    run_stage
+			    USE_QCOW2=1
+			else
+			    run_stage
+			fi
 		fi
 	fi
 done
@@ -253,6 +311,10 @@ if [ -x postrun.sh ]; then
 	cd "${BASE_DIR}"
 	./postrun.sh
 	log "End postrun.sh"
+fi
+
+if [ "${USE_QCOW2}" = "1" ]; then
+    unload_qimage
 fi
 
 log "End ${BASE_DIR}"
